+++ sampledb/logic/schemas/validate.py
@@ -555,11 +555,19 @@ def _validate_object_reference(instance: dict, schema: dict, path: typing.List[s
         object = objects.get_object(object_id=instance['object_id'])
     except ObjectDoesNotExistError:
         raise ValidationError('object does not exist', path)
-    if 'action_id' in schema and isinstance(schema['action_id'], int) and object.action_id != schema['action_id']:
-        raise ValidationError('object has wrong action', path)
-    if 'action_type_id' in schema and isinstance(schema['action_type_id'], int):
-        action = actions.get_action(object.action_id)
-        if action.type_id != schema['action_type_id']:
+    if 'action_id' in schema:
+        if type(schema['action_id']) == int:
+            valid_action_ids = [schema['action_id']]
+        else:
+            valid_action_ids = schema['action_id']
+        if valid_action_ids is not None and object.action_id not in valid_action_ids:
+            raise ValidationError('object has wrong action', path)
+    if 'action_type_id' in schema:
+        if type(schema['action_type_id']) == int:
+            valid_action_type_ids = [schema['action_type_id']]
+        else:
+            valid_action_type_ids = schema['action_type_id']
+        if valid_action_type_ids is not None and actions.get_action(object.action_id).type_id not in valid_action_type_ids:
             raise ValidationError('object has wrong action type', path)